refactor (v-date-picker): extracted common code to mixin
