+++ src/components/VDatePicker/VDatePickerMonthTable.js
@@ -4,7 +4,6 @@ import DatePickerTable from './mixins/date-picker-table'
 
 // Utils
 import { pad, createNativeLocaleFormatter } from './util'
-import isValueAllowed from '../../util/isValueAllowed'
 
 export default {
   name: 'v-date-picker-month-table',
@@ -24,41 +23,6 @@ export default {
     calculateTableDate (delta) {
       return `${parseInt(this.tableDate, 10) + Math.sign(delta || 1)}`
     },
-    genMonthButtonClasses (month) {
-      const isSelected = this.selectedYear === this.displayedYear && this.selectedMonth === month
-      const isCurrent = this.currentYear === this.displayedYear && this.currentMonth === month
-      const isDisabled = !isValueAllowed(`${this.displayedYear}-${pad(month + 1)}`, this.allowedDates)
-      const classes = {
-        'btn--flat': !isSelected,
-        'btn--active btn--depressed': isSelected,
-        'btn--outline': isCurrent && !isSelected,
-        'btn--disabled': isDisabled
-      }
-
-      return isSelected
-        ? this.addBackgroundColorClassChecks(classes)
-        : isCurrent
-          ? this.addTextColorClassChecks(classes)
-          : classes
-    },
-    genMonthButton (month) {
-      const value = `${this.displayedYear}-${pad(month + 1)}`
-      const isDisabled = !isValueAllowed(value, this.allowedDates)
-      const btnContent = this.$createElement('div', {
-        staticClass: 'btn__content'
-      }, [
-        this.formatter(value)
-      ])
-
-      return this.$createElement('button', {
-        staticClass: 'btn',
-        'class': this.genMonthButtonClasses(month),
-        attrs: { type: 'button' },
-        on: isDisabled ? undefined : {
-          click: () => this.$emit('input', `${value}`)
-        }
-      }, [btnContent])
-    },
     genTBody () {
       const children = []
       const cols = Array(3).fill(null)
@@ -67,7 +31,11 @@ export default {
       for (let row = 0; row < rows; row++) {
         const tds = cols.map((_, col) => {
           const month = row * cols.length + col
-          return this.$createElement('td', { key: month }, [this.genMonthButton(month)])
+          return this.$createElement('td', {
+            key: month
+          }, [
+            this.genButton(`${this.displayedYear}-${pad(month + 1)}`, false)
+          ])
         })
 
         children.push(this.$createElement('tr', {
